requirements.txt
numpy
rospy
cv-bridge-3
opencv-python
tensorflow
tensorflow-hub